requirements.txt
Django==1.10.4
django-cors-headers==1.3.1
django-polymorphic==1.0.2
django-storages==1.5.1
djangorestframework==3.5.3
optional-django==0.3.0
Pillow==3.4.2
react==3.0.1
requests==2.12.3
